Fix CodeClimate issue
