+++ app/javascript/widget/store/modules/conversation/index.js
@@ -1,0 +1,23 @@
+import { getters } from './getters';
+import { actions } from './actions';
+import { mutations } from './mutations';
+
+const state = {
+  conversations: {},
+  meta: {
+    userLastSeenAt: undefined,
+  },
+  uiFlags: {
+    allMessagesLoaded: false,
+    isFetchingList: false,
+    isAgentTyping: false,
+  },
+};
+
+export default {
+  namespaced: true,
+  state,
+  getters,
+  actions,
+  mutations,
+};
